- Summary: Fix issues of dynamic parameter adapter
-Bug (s) Resolved:

- Description:
Fix the cleaning up and matching logic for DynamicFilterParameter in ResultSetCriteriaAdapter.
Update test cases for the change.

- Tests Description:
Auto Test

- Files Edited:

- Files Added:

- Files Deleted:

- Notes to Build Team:

- Notes to Developers:

- Notes to QA:

- Notes to Documentation:

+++ model/org.eclipse.birt.report.model.adapter.oda/src/org/eclipse/birt/report/model/adapter/oda/impl/ResultSetCriteriaAdapter.java
@@ -82,11 +82,6 @@ public class ResultSetCriteriaAdapter
 	private final DynamicFilterParameterAdapter paramAdapter;
 
 	/**
-	 * Constant to separate date set name and column name.
-	 */
-	private final static String SEPERATOR = ":"; //$NON-NLS-1$
-
-	/**
 	 * Prefix constant for custom expression.
 	 */
 	private final static String CUSTOM_PREFIX = "#"; //$NON-NLS-1$
@@ -97,7 +92,11 @@ public class ResultSetCriteriaAdapter
 	private final static String DYNAMIC_PREFIX = "!"; //$NON-NLS-1$
 
 	/**
-	 * 
+	 * Constant for invalid dynamic expression.
+	 */
+	private final static String INVALID_FILTER = "^"; //$NON-NLS-1$
+	/**
+	 * Design factory
 	 */
 
 	protected final IODADesignFactory designFactory;
@@ -116,8 +115,7 @@ public class ResultSetCriteriaAdapter
 	{
 		this.setHandle = setHandle;
 		this.setDesign = setDesign;
-		this.paramAdapter = new DynamicFilterParameterAdapter( setHandle,
-				setDesign );
+		this.paramAdapter = new DynamicFilterParameterAdapter( setHandle, setDesign );
 		designFactory = ODADesignFactory.getFactory( );
 	}
 
@@ -177,8 +175,8 @@ public class ResultSetCriteriaAdapter
 
 		int count = 0;
 		FilterExpression filterExpr = null;
-		for ( Iterator<FilterConditionHandle> iter = setHandle
-				.filtersIterator( ); iter.hasNext( ); )
+		for ( Iterator<FilterConditionHandle> iter = setHandle.filtersIterator( ); iter
+				.hasNext( ); )
 		{
 			FilterConditionHandle filterHandle = iter.next( );
 			FilterExpression filter = createOdaFilterExpression( filterHandle );
@@ -315,10 +313,8 @@ public class ResultSetCriteriaAdapter
 			SortKey key = list.get( i );
 			SortHint sortHint = StructureFactory.createSortHint( );
 
-			sortHint.setProperty( SortHint.COLUMN_NAME_MEMBER, key
-					.getColumnName( ) );
-			sortHint.setProperty( SortHint.POSITION_MEMBER, key
-					.getColumnPosition( ) );
+			sortHint.setProperty( SortHint.COLUMN_NAME_MEMBER, key.getColumnName( ) );
+			sortHint.setProperty( SortHint.POSITION_MEMBER, key.getColumnPosition( ) );
 			sortHint.setProperty( SortHint.IS_OPTIONAL_MEMBER, key.isOptional( ) );
 
 			SortDirectionType sortType = key.getSortDirection( );
@@ -385,21 +381,20 @@ public class ResultSetCriteriaAdapter
 
 		filterExpression = criteria.getFilterSpecification( );
 
-		if ( filterExpression == null )
-		{
-			return;
-		}
 		Map<String, Filter> filterExprMap = buildFilterExpressionMap( filterExpression );
 
 		// clears up old filter conditions and finds parameters to refresh
 		cleanUpROMFilterCondition( filterExprMap );
 
-		// update exists filter conditions
-		updateExistingROMFilterConditions( filterExprMap );
-		// sets new filter conditions
-		createROMFilterConditions( filterExprMap );
+		if ( filterExpression != null )
+		{
+			// update exists filter conditions
+			updateExistingROMFilterConditions( filterExprMap );
+			// sets new filter conditions
+			createROMFilterConditions( filterExprMap );
 
-		filterExprMap.clear( );
+			filterExprMap.clear( );
+		}
 		filterExprMap = null;
 	}
 
@@ -410,8 +405,7 @@ public class ResultSetCriteriaAdapter
 	 *            the filter expression
 	 * @return the map containing filter expression to convert
 	 */
-	private Map<String, Filter> buildFilterExpressionMap(
-			FilterExpression filterExpr )
+	private Map<String, Filter> buildFilterExpressionMap( FilterExpression filterExpr )
 	{
 		HashMap<String, Filter> filterExpressions = new LinkedHashMap<String, Filter>( );
 		if ( filterExpr != null )
@@ -422,11 +416,9 @@ public class ResultSetCriteriaAdapter
 				if ( compositeFilterExp instanceof AndExpression )
 				{
 					// Convert and expression only now.
-					for ( FilterExpression child : compositeFilterExp
-							.getChildren( ) )
+					for ( FilterExpression child : compositeFilterExp.getChildren( ) )
 					{
-						filterExpressions
-								.putAll( buildFilterExpressionMap( child ) );
+						filterExpressions.putAll( buildFilterExpressionMap( child ) );
 					}
 				}
 			}
@@ -480,23 +472,17 @@ public class ResultSetCriteriaAdapter
 		{
 			CustomFilter customFilter = (CustomFilter) filter;
 			if ( !StringUtil.isBlank( customFilter.getColumnExpr( ) ) )
-			{
 				key = CUSTOM_PREFIX + customFilter.customFilterExpr.toString( );
-			}
 		}
 		else if ( filter instanceof DynamicFilter )
 		{
 			String columnExpr = ( (DynamicFilter) filter ).getColumnExpr( );
 			if ( !StringUtil.isBlank( columnExpr ) )
-			{
-				key = DYNAMIC_PREFIX + setHandle.getName( ) + SEPERATOR
-						+ columnExpr;
-			}
+				key = DYNAMIC_PREFIX + columnExpr;
 		}
 		else
-		{
 			assert false;
-		}
+
 		return key;
 	}
 
@@ -511,21 +497,22 @@ public class ResultSetCriteriaAdapter
 	private String getMapKey( FilterConditionHandle filterConditionHandle )
 	{
 		String key = null;
-		if ( !StringUtil.isBlank( filterConditionHandle
-				.getDynamicFilterParameter( ) ) )
+		String dynamicFilterParameter = filterConditionHandle
+				.getDynamicFilterParameter( );
+		if ( !StringUtil.isBlank( dynamicFilterParameter ) )
 		{
 			ParameterHandle parameterHandle = setHandle.getModuleHandle( )
-					.findParameter(
-							filterConditionHandle.getDynamicFilterParameter( ) );
+					.findParameter( dynamicFilterParameter );
 			if ( parameterHandle instanceof DynamicFilterParameterHandle )
 			{
-				DynamicFilterParameterHandle dynamicFilterParamHandle = (DynamicFilterParameterHandle) parameterHandle;
 				key = DYNAMIC_PREFIX
-						+ dynamicFilterParamHandle.getDataSetName( )
-						+ SEPERATOR
 						+ ExpressionUtil
-								.createDataSetRowExpression( dynamicFilterParamHandle
+								.createDataSetRowExpression( ( (DynamicFilterParameterHandle) parameterHandle )
 										.getColumn( ) );
+			}
+			else
+			{	// Cannot find the parameter
+				key = INVALID_FILTER;
 			}
 		}
 		return key;
@@ -538,8 +525,8 @@ public class ResultSetCriteriaAdapter
 	 *            the map containing the filters to update
 	 * @throws SemanticException
 	 */
-	private void updateExistingROMFilterConditions(
-			Map<String, Filter> filterMap ) throws SemanticException
+	private void updateExistingROMFilterConditions( Map<String, Filter> filterMap )
+			throws SemanticException
 	{
 		for ( Iterator iter = setHandle.filtersIterator( ); iter.hasNext( ); )
 		{
@@ -550,8 +537,7 @@ public class ResultSetCriteriaAdapter
 			{
 				DynamicFilterParameterHandle dynamicFilterParamHandle = (DynamicFilterParameterHandle) setHandle
 						.getModuleHandle( ).findParameter(
-								filterConditionHandle
-										.getDynamicFilterParameter( ) );
+								filterConditionHandle.getDynamicFilterParameter( ) );
 				updateDynamicFilterCondition( filterConditionHandle,
 						(DynamicFilter) filterMap.get( key ),
 						dynamicFilterParamHandle );
@@ -577,8 +563,7 @@ public class ResultSetCriteriaAdapter
 	{
 		for ( Filter filter : filterMap.values( ) )
 		{
-			FilterCondition filterCondition = StructureFactory
-					.createFilterCond( );
+			FilterCondition filterCondition = StructureFactory.createFilterCond( );
 			filterCondition.setExpr( filter.getColumnExpr( ) );
 
 			// in default, to make sure the pushdown is true.
@@ -590,8 +575,7 @@ public class ResultSetCriteriaAdapter
 			if ( filter instanceof CustomFilter )
 			{
 				CustomFilterExpression customFilterExp = ( (CustomFilter) filter ).customFilterExpr;
-				updateCustomFilterCondition( filterConditionHandle,
-						customFilterExp );
+				updateCustomFilterCondition( filterConditionHandle, customFilterExp );
 			}
 			else if ( filter instanceof DynamicFilter )
 			{
@@ -600,25 +584,19 @@ public class ResultSetCriteriaAdapter
 				DynamicFilterParameterHandle dynamicFilterParamHandle = setHandle
 						.getModuleHandle( ).getElementFactory( )
 						.newDynamicFilterParameter( null );
-				dynamicFilterParamHandle.setDataSetName( setHandle.getName( ) );
-				dynamicFilterParamHandle.setColumn( dynamicFilter
-						.getColumnName( ) );
+				dynamicFilterParamHandle.setColumn( dynamicFilter.getColumnName( ) );
 				Integer nativeDataType = dynamicFilter.getNativeDataType( );
 				if ( nativeDataType != null )
 				{
 					dynamicFilterParamHandle.setNativeDataType( nativeDataType );
 					try
 					{
-						dynamicFilterParamHandle
-								.setDataType( NativeDataTypeUtil
-										.getUpdatedDataType(
-												setDesign
-														.getOdaExtensionDataSourceId( ),
-												setDesign
-														.getOdaExtensionDataSetId( ),
-												nativeDataType,
-												null,
-												DesignChoiceConstants.CHOICE_PARAM_TYPE ) );
+						dynamicFilterParamHandle.setDataType( NativeDataTypeUtil
+								.getUpdatedDataType( setDesign
+										.getOdaExtensionDataSourceId( ), setDesign
+										.getOdaExtensionDataSetId( ),
+										nativeDataType, null,
+										DesignChoiceConstants.CHOICE_PARAM_TYPE ) );
 					}
 					catch ( BirtException e )
 					{
@@ -632,8 +610,8 @@ public class ResultSetCriteriaAdapter
 						.setDynamicFilterParameter( dynamicFilterParamHandle
 								.getName( ) );
 				// updates the dynamic filter parameter
-				updateDynamicFilterCondition( filterConditionHandle,
-						dynamicFilter, dynamicFilterParamHandle );
+				updateDynamicFilterCondition( filterConditionHandle, dynamicFilter,
+						dynamicFilterParamHandle );
 			}
 		}
 	}
@@ -698,10 +676,8 @@ public class ResultSetCriteriaAdapter
 		ArrayList<FilterCondition> dropList = new ArrayList<FilterCondition>( );
 		for ( Iterator iter = setHandle.filtersIterator( ); iter.hasNext( ); )
 		{
-			FilterConditionHandle filterHandle = (FilterConditionHandle) iter
-					.next( );
-			String dynamicParameterName = filterHandle
-					.getDynamicFilterParameter( );
+			FilterConditionHandle filterHandle = (FilterConditionHandle) iter.next( );
+			String dynamicParameterName = filterHandle.getDynamicFilterParameter( );
 			String key = getMapKey( filterHandle );
 			// Check if contains such filter.
 			if ( key != null && !filterMap.containsKey( key ) )
@@ -709,10 +685,10 @@ public class ResultSetCriteriaAdapter
 				// Remove the filter condition which is not contained.
 				if ( !StringUtil.isBlank( dynamicParameterName ) )
 				{
-					ParameterHandle parameterHandle = setHandle
-							.getModuleHandle( ).findParameter(
-									dynamicParameterName );
-					parameterHandle.drop( );
+					ParameterHandle parameterHandle = setHandle.getModuleHandle( )
+							.findParameter( dynamicParameterName );
+					if ( parameterHandle != null )
+						parameterHandle.drop( );
 				}
 				dropList.add( (FilterCondition) filterHandle.getStructure( ) );
 			}
@@ -746,8 +722,7 @@ public class ResultSetCriteriaAdapter
 			CustomFilterExpression customFilterExpr = designFactory
 					.createCustomFilterExpression( );
 
-			ExpressionVariable variable = designFactory
-					.createExpressionVariable( );
+			ExpressionVariable variable = designFactory.createExpressionVariable( );
 			variable.setIdentifier( filterHandle.getExpr( ) );
 			customFilterExpr.setContextVariable( variable );
 
@@ -790,8 +765,7 @@ public class ResultSetCriteriaAdapter
 				paramAdapter.updateODADynamicFilter( paramDefn, defaultType,
 						dynamicParamHandle );
 
-				paramDefn.getAttributes( ).setName(
-						dynamicParamHandle.getColumn( ) );
+				paramDefn.getAttributes( ).setName( dynamicParamHandle.getColumn( ) );
 				if ( dynamicParamHandle
 						.getProperty( IDynamicFilterParameterModel.NATIVE_DATA_TYPE_PROP ) != null )
 				{
@@ -897,8 +871,7 @@ public class ResultSetCriteriaAdapter
 		 */
 		public String getColumnName( )
 		{
-			ParameterDefinition paramDefn = exprParamDefn
-					.getDynamicInputParameter( );
+			ParameterDefinition paramDefn = exprParamDefn.getDynamicInputParameter( );
 			if ( paramDefn != null && paramDefn.getAttributes( ) != null )
 			{
 				return paramDefn.getAttributes( ).getName( );
@@ -931,8 +904,7 @@ public class ResultSetCriteriaAdapter
 
 		public Integer getNativeDataType( )
 		{
-			ParameterDefinition paramDefn = exprParamDefn
-					.getDynamicInputParameter( );
+			ParameterDefinition paramDefn = exprParamDefn.getDynamicInputParameter( );
 			if ( paramDefn != null && paramDefn.getAttributes( ) != null )
 			{
 				return paramDefn.getAttributes( ).getNativeDataTypeCode( );
